+++ sdk/data/azappconfig/internal/synctoken/cache_test.go
@@ -7,6 +7,7 @@
 package synctoken
 
 import (
+	"strings"
 	"testing"
 
 	"github.com/stretchr/testify/require"
@@ -35,5 +36,9 @@ func TestCache(t *testing.T) {
 
 	require.NoError(t, stk.Set("id2=some;sn=2"))
 	f = stk.Get()
-	require.EqualValues(t, "id=val2,id2=some", f)
+	// NOTE: Get() ranges over a map and the order is non-deterministic so we can't perform a simple equals check
+	//require.EqualValues(t, "id=val2,id2=some", f)
+	require.Contains(t, f, "id=val2")
+	require.Contains(t, f, "id2=some")
+	require.EqualValues(t, 1, strings.Count(f, ","))
 }